A simple graph class for dependencies
It is the full graph. I.e. all binaries and their dependencies sit in
their own nodes. Then there should be some instantiation of this in the
"dependency soup" directories. I.e. resolve node conflicts and dump them
into flat sets etc.

Also, this graph is just for 1 binary.

Probably it's better to just create the final graph on the fly, without
the full one as a preliminary step. Keep a set with all nodes, or node
names, check for collisions, resolve them in the graph with a new node.
Something like that.

I.e. there should be something like:
* a file with node definitions - construct a graph out of it - construct
  the "soup" out of that, also find the binaries for the nodes in the
  "store" directories, check their RPATH
* that file describes the environment
* extract the definition graph from existing binaries and dump it into
  the definitions file, also copy the binaries and convert them into the
  store format

I should be able to make a store out of existing binaries, and
definitions for my environment. Then check the hashes of the files, the
hash of the environment. Something around that.

	modified:   all_deps.py
	new file:   dep_node.py