And it pulls the dependencies from the store dir
It pulls them, and is supposed it check the rules. But I have not tested
them. It looks up in the rules file poorly now:

> TODO this won't work! the definitions are hashed - they won't match loose versions etc!

The rules file is supposed to describe a dictionary with:

    {
    DepDefinition: {filename: DepDefinition, ...}
    ...
    }

The look up by DepDefinition key is a faulty thing: it matches
DepDefinitions precisely, when the semantics is to match them loosely.
If the versions definition is loose, you match something within the
range. The hash tags are matched against a set, etc.

Maybe, it's enough to change it to:

    {
    binary_filename: {DepDefinition: {filename: DepDefinition, ...}, ...}
    ...
    }

So, you can have _different_ rules for the _same_ binary. Of course, at
the top level you can have only one binary_filename. But for deeper
dependencies, there can be many.

But this is poor. Looking up a definition is painful: look up the name,
then loop over definitions to find a match. That's useless actions.

It's better to implement some custom mapping class, with the semantic
definition look up.

	modified:   all_deps.py
	modified:   store_files.py